Working on the upgrade system
Added GetName() metod to the Item class so that it is possible to retrieve the items name.
Made the Utility class into an abstract class and added an implementation for the GetUpgradeOptions() method.
Made the YoyoController into a PermanentProjectileWeapon instead of ProjectileWeapon.
Added an implementation for the GetUpgradeOptions() in the TetheringWeapon abstract class.
Made it so that the Player class calls the StartUpgradeSystem() method whenever the player levels up.
Added a panelList in order to be able to remove already occupied panels.
Added the method InitializeUpgradeOptions() in order to find what the player is able to upgrade.
Added the method RemoveUpgradeOption() in order to remove any option that has already been selected to be presented.
Added the method ChooseUpgradeType() in order to determine what kind of option the player should receive.
Added the method GetUpgradableItems() in order to get all the items that the player is able to upgrade.
Added the method ChooseRandomPanel() in order to randomize which option appears on which panel.
Added the method SetPanelText() in order to set the text of the panel to the thing that it will award the player with.
Added the method RemovePanelAsOption() in order to effectively mark the panel as occupied.
Added the method SetPanelMethod() in order to send the method that will call the appropriate upgrade method in each Item class to the panel button.
Added some of the implementation for the StartUpgradeSystem() method that will go through all of the steps in order to set it up correctly.
Added the PerformRandomizedAction() method that will call the appropriate method which will upgrade an ability.
Added the DetermineTypeOptions() method which will figure out what kind of options the player is able to receive based on what they currently have.